% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAT.R
\name{CCAT}
\alias{CCAT}
\title{Correlation of Connectome And Transcriptome}
\usage{
CCAT(
  Integration.l = NULL,
  data.m = NULL,
  ppiA.m = NULL,
  log_trans = FALSE,
  parallelMode = FALSE,
  mcores = 1,
  subsamplesize = 1000
)
}
\arguments{
\item{Integration.l}{A list object from \code{DoIntegPPI} function.}

\item{data.m}{A scRNA-Seq data matrix with rows labeling genes and columns 
labeling single cells. And it can be either a log-transformed data
matrix with minimal value around 0.1 (recommended), or an 
nonlog-transformed data matrix with minimal value 0.}

\item{ppiA.m}{The adjacency matrix of a user-given PPI network with rownames and 
colnames labeling genes (same gene identifier as in \code{exp.m})}

\item{log_trans}{A logical. Whether to do log-transformation on the input data
matrix or not. Default is FALSE}

\item{parallelMode}{A logical. Indicating whether or not to run CCAT in parallel. 
It will be disable if datasets are < 5,000 cells. Parallel mode 
uses a subsampling approach to reduce runtime. Default is FALSE}

\item{mcores}{A integer. Indicating the number of cores to use when parallelMode = TRUE}

\item{subsamplesize}{A integer. Indicating the number of cells to subsample when parallelMode = TRUE}
}
\value{
A list incorporates the input list and CCAT velues or CCAT values 
itself, depending on the input object(s):

CCAT
The estimated signaling entropy rate using Pearson correlation coefficient
}
\description{
This function leverages Pearson correlation between gene expresion level
and gene connectome derived from PPI network to fastly estimate signaling
entropy rate.
}
\examples{
### load example data & network matrix
data(Example.m)
data(net13Jun12.m)

### integrate expr matrix and PPI network
Integration.l <- DoIntegPPI(exp.m = Example.m, ppiA.m = net13Jun12.m)

### estimate SR with PCC
### get it with the integration list
Integration.l <- CCAT(Integration.l)

### or get CCAT directly from data matrix
CCAT.v <- CCAT(data.m = Example.m, ppiA.m = net13Jun12.m)


}
\references{
Chen, Weiyan, et al.
\emph{Single-cell landscape in mammary epithelium reveals 
bipotent-like cells associated with breast cancer risk 
and outcome.}
Communications Biology 2 (2019): 306.
doi:\href{https://doi.org/10.1038/s42003-019-0554-8}{
10.1038/s42003-019-0554-8}.

Teschendorff Andrew E., Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff Andrew E., Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
