% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSRana.R
\name{CompSRana}
\alias{CompSRana}
\title{Computes the signaling entropy of given gene expression profiles 
over a given connected network}
\usage{
CompSRana(Integration.l, local = FALSE, mc.cores = 1)
}
\arguments{
\item{Integration.l}{A list object from \code{DoIntegPPI} function.}

\item{local}{A logical (default is FALSE). If TRUE, function computes the normalized 
local signaling entropies of each gene in the network.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will 
be run simultaneously. The option is initialized from environment variable 
MC_CORES if set. Must be at least one, and parallelization requires at 
least two cores.}
}
\value{
A list incorporates the input list and four new elements:

SR
The global signaling entropy rate. It is normalized by the 
maximum rate, hence a value between 0 and 1

inv
The stationary distribution of every sample

s
The unnormlised local entropies of each gene in every cell

ns
The normalised local entropies of each gene, so that each value is 
between 0 and 1
}
\description{
This is the main user function for computing signaling entropy of 
single cells. It takes as input the gene expression profile of 
single cells and the adjacency matrix of a connected network. These 
inputs will be typically the output of the \code{DoIntegPPI} function.
}
\examples{
### define a small network
ppiA.m <- matrix(0,nrow=10,ncol=10);
ppiA.m[1,] <- c(0,1,1,1,1);
for(r in 2:nrow(ppiA.m)){
  ppiA.m[r,1] <- 1;
}
rownames(ppiA.m) <- paste("G",1:10,sep="");
colnames(ppiA.m) <- paste("G",1:10,sep="");

### define a positively valued expression matrix (20 genes x 10 samples)
exp.m <- matrix(rpois(20*10,8),nrow=20,ncol=10);
colnames(exp.m) <- paste("S",1:10,sep="");
rownames(exp.m) <- paste("G",1:20,sep="");

### integrate data and network
Integration.l <- DoIntegPPI(exp.m, ppiA.m);

### compute SR values
Integration.l <- CompSRana(Integration.l);

### output global signaling entropy
print(Integration.l$SR);

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff AE, Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
