% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoDiffusionMap.R
\name{DoDiffusionMap}
\alias{DoDiffusionMap}
\title{Construct diffusion map and visualization}
\usage{
DoDiffusionMap(Integration.l, mean_gap = 1, sd_gap = 1,
  root = c("cell", "state"), dim = c(1, 2), k = 30, phi = 20,
  theta = 30, color_by = c("SR", "DPT"), PLOT = FALSE, PDF = FALSE)
}
\arguments{
\item{Integration.l}{Typically, it is the output from \code{InferPotency} function}

\item{mean_gap}{The threshold for HVG gene selection based on the mean expression
level. Default value is 1}

\item{sd_gap}{The threshold for HVG gene selection based on the expression
level's standard deviation. Default value is 1}

\item{root}{Define the root of the trajectory. \code{cell} indicates the
root is the cell with the highest SR value among all cells;
\code{state} indicates the root is the cell with the highest 
SR value within the highest inferred potency state}

\item{dim}{A numeric vector. Dimension used in calculating diffusion 
components and the following plot generation.
Default is c(1, 2)}

\item{k}{Number of nearest neighbors to consider. Default is 30}

\item{phi}{The angles defining the viewing direction. \code{phi} 
gives the colatitude. Default is 20}

\item{theta}{The angles defining the viewing direction. \code{theta} 
gives the azimuthal direction. Default is 30}

\item{color_by}{Indicating the variable used for coloring. \code{SR}
means the plot is colored by SR values. \code{DPT}
means the plot is colored by diffusion pseudotime}

\item{PLOT}{A logical. Decides whether to generate the plot
or not, default is FALSE}

\item{PDF}{A logical. Output figure via pdf file or not, default is FALSE}
}
\value{
An updated list objects, add three elements:

DM
A diffusion map object

DMEigen
The diffusion components of the diffusion map

DPT
The diffusion pseudotime generated from the diffusion map
}
\description{
This function constructs diffusion map with highly variable
genes and visualize the result in 2D or 3D plots
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.
}
