% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoLandSCENT.R
\name{DoLandSCENT}
\alias{DoLandSCENT}
\title{Runs the LandSCENT algorithm}
\usage{
DoLandSCENT(exp.m, ppiA.m, mc.cores = 1, pheno.v = NULL,
  reducedMatrix = NULL, reduceDim = TRUE, PLOT = TRUE, PDF = TRUE)
}
\arguments{
\item{exp.m}{Can be three kinds of input.
One is a scRNA-Seq data matrix with rows labeling genes and columns 
labeling single cells, which is not log-transformed (can be FPKM, TPM etc.).
And rownames must be provided and must be the same with the 
rownames of \code{ppiA.m}.
The other kinds of two input can be either a "SingleCellExperiment"
class object or a "CellDataSet" class object}

\item{ppiA.m}{The adjacency matrix of a user-given PPI network with rownames and 
colnames labeling genes (same gene identifier as in \code{exp.m}).}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will 
be run simultaneously. The option is initialized from environment variable 
MC_CORES if set. Must be at least one, and parallelization requires at 
least two cores.}

\item{pheno.v}{A phenotype vector for the single cells, of same length and order as the 
columns of \code{exp.m}.
Function can also automatically extract phenotype information
from your original sce/cds data, please store the phenotype information
as name of \code{phenoInfo}.}

\item{reducedMatrix}{The previous reduced dimension matrix, with rows lalabeling cells and two
colums labeling reduced dimensions.(Optional, only used when 
\code{reduceDim} set to \code{FALSE})}

\item{reduceDim}{A logical, do deminsion reduction or not before generate the plots.
Default is TRUE.}

\item{PLOT}{A logical. Decides whether to generate (default) the landSR figure 
and CellSR figure or not.}

\item{PDF}{The figure file output format, via pdf (TRUE) file or not, default is TRUE.}
}
\value{

}
\description{
Main user function implement LandSCENT. This function is the typical 
workflow of the whole package for you to easily use.
}
