% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferLandmark.R
\name{InferLandmark}
\alias{InferLandmark}
\title{Infer co-expression landmarks from cells' SR values}
\usage{
InferLandmark(Integration.l, pheno.v = NULL, pctG = 0.01,
  Component_use = NULL, reduceMethod = c("PCA", "tSNE"),
  clusterMethod = c("PAM", "dbscan"), k_pam = 9, eps_dbscan = 10,
  minPts_dbscan = 5, pctLM = 0.05, pcorTH = 0.1)
}
\arguments{
\item{Integration.l}{A list object from \code{InferPotency} function.}

\item{pheno.v}{A phenotype vector for the single cells, of same length and order as the 
columns of \code{Integration.l$expMC}.
Function can also automatically extract phenotype information
from your original sce/cds data, please store the phenotype information
under the name of \code{phenoInfo}.}

\item{pctG}{A numeric. Percentage of all genes in \code{Integration.l$expMC} 
to select from each principal component in an SVD/PCA 
of \code{Integration.l$expMC}. The union set of all selected genes 
is then used for clustering. Default value is 0.01.}

\item{Component_use}{A numeric. Specify the number of principal components in the PCA to use
in the downstream analysis. Default value is NULL.}

\item{reduceMethod}{A character, either "PCA" or "tSNE". Indicates using PCA or
tSNE method to do dimension reduction.}

\item{clusterMethod}{A character, either "PAM" or "dbscan". Indicates using dbscan or
PAM method to do clustering.}

\item{k_pam}{Only used when \code{clusterMethod} is set to be "PAM".
Maximum number of co-expression clusters, when performing clustering.
Default value is 9.}

\item{eps_dbscan}{Only used when \code{clusterMethod} is set to be "dbscan". 
Size of the epsilon neighborhood. Default is 10.}

\item{minPts_dbscan}{Only used when \code{clusterMethod} is set to be "dbscan".
Number of minimum points in the eps region (for core points).
Default is 5 points.}

\item{pctLM}{Percentage of total number of single cells to allow as a minimum size for 
selecting interesting landmarks i.e. potency-coexpression clusters of single 
cells. Default value is 0.05.}

\item{pcorTH}{Threshold for calling significant partial correlations. Default value is 0.1.
Usually, single-cell experiments profile large number of cells, so 0.1 is a 
sensible threshold.}
}
\value{
Integration.l
A list incorporates the input list with a new list named 
\code{InferLandmark.l}.

InferLandmark.l
A list contains twelve objects:

cl
The co-expression clustering index for each single cell

pscl
The potency coexpression clustering label for each single cell

distPSCL
The distribution of single cell numbers per potency state and coexpression 
cluster

medLM
A matrix of medoids of gene expression for the selected landmarks

srPSCL
The average signaling entropy of single cells in each potency coexpression 
cluster

srLM
The average signaling entropy of single cells in each landmark

distPHLM
Table giving the distribution of single cell numbers per phenotype and 
landmark

cellLM
Nearest landmark for each single cell

cellLM2
A vector specifying the nearest and next-nearest landmark for each single 
cell

adj
Weighted adjacency matrix between landmarks with entries giving the number 
of single cells mapping closest to the two landmarks

pcorLM
Partial correlation matrix of landmarks as estimated from the expression 
medoids

netLM
Adjacency matrix of landmarks specifying which partial correlations are 
significant

selectGene
Selected a group of genes for internal clustering
}
\description{
This function identifies potency-coexpression clusters of single cells, 
called landmarks, and finally infers the dependencies of these landmarks 
which can aid in recontructing lineage trajectories in time course 
or development associated scRNA-Seq experiments.
}
\examples{
data(Example.m)
data(net13Jun12.m)
Integration.l <- DoIntegPPI(exp.m = Example.m[, c(1:58,61:84,86:98,100)], ppiA.m = net13Jun12.m)
data(SR.v)
Integration.l$SR <- SR.v
InferPotency.o <- InferPotency(Integration.l)
InferLandmark.o <- InferLandmark(InferPotency.o)

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff AE, Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
