% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferPotency.R
\name{InferPotency}
\alias{InferPotency}
\title{Infer Distinct potency states from cells' SR values}
\usage{
InferPotency(Integration.l, pheno.v = NULL, diffvar = TRUE,
  maxPS = 5)
}
\arguments{
\item{Integration.l}{A list object from \code{CompSRana} function.}

\item{pheno.v}{A phenotype vector for the single cells, of same length and order as the 
columns of \code{Integration.l$expMC}.
Function can also automatically extract phenotype information
from your original sce/cds data, please store the phenotype information
under the name of \code{phenoInfo}}

\item{diffvar}{A logical. Default is TRUE.
Specifies whether the Gaussian mixture model to be fit assumes components 
to have different (default) or equal variance.
In the latter case, use *modelNames = c("E")*.}

\item{maxPS}{Maximum number of potency states, when inferring discrete potency 
states of single cells. Default value is 5.}
}
\value{
Integration.l
A list incorporates the input list with some new elements.

Integration.l$potencyState
Inferred discrete potency states for each single cell. It is indexed so 
that the index increases as the signaling entropy of the state decreases

Integration.l$distPSPH
If phenotype information provided, it will be a table giving the 
distribution of single-cells across potency states and 
phenotypes

Integration.l$prob
Table giving the probabilities of each potency state per phenotype value

Integration.l$hetPS
The normalised Shannon Index of potency per phenotype value
}
\description{
This function infers the discrete potency states of single cells and 
its distribution across the single cell population.
}
\examples{
data(Example.m)
data(net13Jun12.m)
Integration.l <- DoIntegPPI(exp.m = Example.m[, c(1:58,61:84,86:98,100)], ppiA.m = net13Jun12.m)
data(SR.v)
Integration.l$SR <- SR.v
InferPotency.o <- InferPotency(Integration.l)

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff AE, Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
