% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandSCENT-package.R
\docType{package}
\name{LandSCENT-package}
\alias{LandSCENT-package}
\title{Landscape Single Cell Entropy}
\description{
\pkg{LandSCENT} (Landscape Single Cell Entropy) is a R-package for the 
analysis of single-cell RNA-Seq data. One important feature of this 
package is the computation of signaling entropy, which allows single 
cells to be ordered according to differentiation potency. \pkg{LandSCENT} 
also integrates cell density with potency distribution to dissect cell types
across all potency states and generates high-quality figures to show this.
}
\details{
\packageIndices{LandSCENT}
\pkg{LandSCENT} will be of interest to those analysing 
single-cell RNA-Sequencing data. A core component of \pkg{LandSCENT} 
is the computation of signaling entropy at the single-cell level
(\code{CompSRana}), allowing cells to be ordered according to
differentiation potency or phenotypic plasticity. It also incorporates
functionality for quantifying intercellular heterogeneity, for identifying
interesting subpopulations of cells that differ in terms of potency or
plasticity, as well as to infer dependencies between single cell
clusters, which for instance can help identify lineage trajectories.
}
\examples{
### see example for CompSRana function for typical workflow

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff AE, Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
\author{
Weiyan Chen & Andrew E Teschendorff
}
