% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net13Jun12.m}
\alias{net13Jun12.m}
\title{Human protein-protein interaction network 13Jun12}
\format{
A matrix with 8434 rows and 8434 columns
}
\usage{
data(net13Jun12.m)
}
\description{
This human protein-protein interaction network is derived from Pathway 
Commons (www.pathwaycommons.org) (version Jun. 2012), which is 
an integrated resource collating together PPIs from several 
distinct sources.
}
\details{
\itemize{
  \item 121268 : Human Gene EntrzID 
  \item 2193 : Human Gene EntrzID 
  \item 79902 : Human Gene EntrzID
  \item ...
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
\keyword{network}
