% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net17Jan16.m}
\alias{net17Jan16.m}
\title{Human protein-protein interaction network 17Jan16}
\format{A matrix with 11751 rows and 11751 columns}
\usage{
data(net17Jan16.m)
}
\description{
This human protein-protein interaction network is derived from Pathway 
Commons (www.pathwaycommons.org) (version Jan. 2016), which is 
an integrated resource collating together PPIs from several 
distinct sources.
}
\details{
\itemize{
  \item 121268 : Human Gene EntrzID
  \item 2193 : Human Gene EntrzID
  \item 79902 : Human Gene EntrzID
  \item ...
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
\keyword{network}
