% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenoscChu.v}
\alias{phenoscChu.v}
\title{Phenotype information for Chu et al. scRNA-seq data}
\format{A character vector includes 1018 elements}
\usage{
data(phenoscChu.v)
}
\description{
Phenotype information for the normalized scRNA-seq data matrix from Chu et al. 2016.
}
\details{
\itemize{
  \item hESC : human embryonic stem cells(n=374)
  \item NPC : ectoderm progenitor(n=173)
  \item DEC : endoderm progenitor(n=138)
  \item EC : mesoderm progenitor cells(n=105)
  \item HFF : human foreskin fibroblasts(n=159)
  \item TB : trophoblasts(n=69)
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Chu Li-Fang, Ning Leng, Jue Zhang, Zhonggang Hou, Daniel Mamott, 
David T. Vereide, Jeea Choi, Christina Kendziorski, Ron Stewart and James A. Thomson
\emph{Single-cell RNA-seq reveals novel regulators of human embryonic 
stem cell differentiation to definitive endoderm.}
Genome biology 17.1 (2016): 173.
doi:\href{https://doi.org/10.1186/s13059-016-1033-x}{
10.1186/s13059-016-1033-x}.
}
\keyword{datasets}
