% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DiffusionMap.R
\name{Plot_DiffusionMap}
\alias{Plot_DiffusionMap}
\title{Visualize diffusion map}
\usage{
Plot_DiffusionMap(
  Integration.l,
  dim = c(1, 2, 3),
  color_by = c("SR", "DPT"),
  TIPs = c(1, 2, 3),
  phi = 20,
  theta = 40,
  bty = "g",
  ...,
  PDF = FALSE
)
}
\arguments{
\item{Integration.l}{Typically, it is the output from \code{DoDiffusionMap} function}

\item{dim}{A numeric vector. Diffusin components order in the plot axes. And 
the sign of evrey entry indicates the direction of component.
Default is c(1, 2, 3)}

\item{color_by}{Indicating the variable used for coloring. \code{SR}
means the plot is colored by SR values. \code{DPT}
means the plot is colored by diffusion pseudotime}

\item{TIPs}{Specifying which trajectory paths should be shown on the 
diffusion maps. Only available in 2D plot and when 
\code{color_by} was set to be \code{DPT}}

\item{phi}{The angles defining the viewing direction. \code{phi} 
gives the colatitude. Default is 20}

\item{theta}{The angles defining the viewing direction. \code{theta} 
gives the azimuthal direction. Default is 40}

\item{bty}{The type of the box, the default \code{g} draws grey 
background with white grid lines.
Note: the bty = "f", "b", "b2", "bl" can also be specified
for this argument}

\item{...}{Additional arguments passed to the 3D plotting methods}

\item{PDF}{A logical. Output figure via pdf file or not, default is FALSE}
}
\value{
A pdf file contains the generated figure or a ggplot object
}
\description{
This function visualizes the diffusion map in 2D or 3D plots
}
\examples{
\dontrun{
data(Example.m)
data(net13Jun12.m)
Integration.l <- DoIntegPPI(exp.m = Example.m[, c(1:58,61:84,86:98,100)], ppiA.m = net13Jun12.m)
data(SR.v)
Integration.l$SR <- SR.v
InferPotency.o <- InferPotency(Integration.l)
DoDM.o <- DoDiffusionMap(InferPotency.o)
Plot_DiffusionMap(DoDM.o)
}

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.
}
