% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_LandSR.R
\name{Plot_LandSR}
\alias{Plot_LandSR}
\title{Plot Landscape with potency states}
\usage{
Plot_LandSR(
  Integration.l,
  coordinates = NULL,
  num_grid = 50,
  phi = 20,
  theta = 20,
  colpersp = NULL,
  colimage = NULL,
  colkeypersp = list(length = 0.2, width = 0.4, shift = 0.15, cex.axis = 0.6, cex.clab
    = 0.65),
  colkeyimage = list(length = 0.2, width = 0.4, shift = -0.15, cex.axis = 0.6, cex.clab
    = 0.65),
  lighting = FALSE,
  lphi = 90,
  bty = "b",
  scale_z = TRUE,
  PDF = TRUE
)
}
\arguments{
\item{Integration.l}{Typically, it is the output from \code{InferLandmark} function}

\item{coordinates}{Optional. The previous reduced dimension coordinates, with rows lalabeling cells 
and two colums labeling reduced dimensions}

\item{num_grid}{Number of grid points in each direction. Can be scalar or a length-2 
integer vector}

\item{phi}{The angles defining the viewing direction. phi gives the colatitude}

\item{theta}{The angles defining the viewing direction. theta gives the 
azimuthal direction}

\item{colpersp}{Color palette to be used for the colvar variable, i.e. specific 
potency state cell density.
If colpersp is NULL (default) and colvar is specified, then a 
red-yellow-blue colorscheme will be used.}

\item{colimage}{Color palette to be used for the colvar variable, i.e. the cell density.
If colimage is NULL (default) and colvar is specified, then a 
red-magenta-white colorscheme will be used.}

\item{colkeypersp}{A logical, or a list (default) with parameters for the color 
key (legend). List parameters should be one of side, plot, 
length, width, dist, shift, addlines, col.clab, cex.clab, 
side.clab, line.clab, adj.clab, font.clab and the axis parameters 
at, labels, tick, line, pos, outer, font, lty, lwd, lwd.ticks, 
col.box, col.axis, col.ticks, hadj, padj, cex.axis, mgp, tck, 
tcl, las. 
The defaults for the parameters are length = 0.2, width = 0.4, 
shift = 0.15, cex.axis = 0.6, cex.clab = 0.65
The default is to draw the color key on side = 4, i.e. in the right 
margin. 
If colkeypersp = NULL then a color key will be added only if 
col is a vector. Setting colkeypersp = list(plot = FALSE) will create 
room for the color key without drawing it. if colkeypersp = FALSE, 
no color key legend will be added.
See more details in ?plot3D::persp3D}

\item{colkeyimage}{A logical, or a list (default) with parameters for the color 
key (legend). List parameters should be one of side, plot, 
length, width, dist, shift, addlines, col.clab, cex.clab, 
side.clab, line.clab, adj.clab, font.clab and the axis parameters 
at, labels, tick, line, pos, outer, font, lty, lwd, lwd.ticks, 
col.box, col.axis, col.ticks, hadj, padj, cex.axis, mgp, tck, 
tcl, las. 
The defaults for the parameters are length = 0.2, width = 0.4, 
shift = -0.15, cex.axis = 0.6, cex.clab = 0.65
The default is to draw the color key on side = 4, i.e. in the right 
margin. 
If colkeyimage = NULL then a color key will be added only if 
col is a vector. Setting colkeyimage = list(plot = FALSE) will create 
room for the color key without drawing it. if colkeyimage = FALSE, 
no color key legend will be added.
See more details in ?plot3D::image3D}

\item{lighting}{A logical. If TRUE, the facets will be illuminated, and colors may 
appear more bright. Default is FALSE}

\item{lphi}{if finite values are specified for lphi, the surface is 
shaded as though it was being illuminated from the direction 
specified by colatitude lphi}

\item{bty}{The type of the box ("b" or "f"), the default ("b") only 
drawing background panels}

\item{scale_z}{A logical, whether to scale the density z value or not}

\item{PDF}{A logical. Output figure via pdf file or not, default is TRUE}
}
\value{
A pdf file contains the generated figures

Integration.l
If coordinates provided, it will be a list object integrates the input
list and coordinates
}
\description{
Density based visualization tool to generate figures that give the cell 
density distribution against potency states
}
\details{
Density based visualization tool to generate figures that give the cell 
density distribution against potency states
}
\examples{
data(tsne.o)
data(potS.v)
data(SR4.v)
potS.v <- 4 - potS.v
InferLandmark.o <- list(potencyState = potS.v, SR = SR4.v, coordinates = tsne.o)

LandSR.o <- Plot_LandSR(InferLandmark.o, PDF = FALSE)

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
