% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotencyInfer.R
\name{PotencyInfer}
\alias{PotencyInfer}
\title{Infer Distinct potency states from cells' SR values}
\usage{
PotencyInfer(Integrataion.l, pheno.v = NULL, mixmod = TRUE,
  maxPS = 5, pctG = 0.01, kmax = 9, reduceMethod = "PAM",
  num_clusters = NULL, epsMax = 10, minPts = 5, pctLM = 0.05,
  pcorTH = 0.1)
}
\arguments{
\item{Integrataion.l}{A list object from \code{CompSRana} function.}

\item{pheno.v}{A phenotype vector for the single cells, of same length and order as the 
columns of \code{Integrataion.l$expMC}.
Function can also automatically extract phenotype information
from your original sce/cds data, please store the phenotype information
as name of \code{phenoInfo}}

\item{mixmod}{A logical. Default is TRUE.
Specifies whether the Gaussian mixture model to be fit assumes components 
to have different (default) or equal variance.
In the latter case, use *mixmod=c("E")*.}

\item{maxPS}{Maximum number of potency states to allow, when inferring discrete potency 
states of single cells. Default value is 5.}

\item{pctG}{Percentage of all genes in \code{Integrataion.l$expMC} to select from
each principal component in an SVD/PCA of \code{Integrataion.l$expMC}.
The union set of all selected genes is then used for clustering. 
Default value is 0.01.}

\item{kmax}{Maximum number of co-expression clusters allowed when performing clustering.
Default value is 9. Larger values are not allowed.}

\item{reduceMethod}{A character, either "tSNE" or "PAM"(default). Indicates using tSNE or
PAM method to do dimension reduction.}

\item{num_clusters}{Only used when data reduced by tSNE, specify the number of clusters.}

\item{epsMax}{Maximum value of scanning when implementing \code{dbscan} to determine
the cluster number.}

\item{minPts}{Number of minimum points in the eps region (for core points).
Default is 5 points.}

\item{pctLM}{Percentage of total number of single cells to allow as a minimum size for 
selecting interesting landmarks i.e. potency-coexpression clusters of single 
cells. Default value is 0.05.}

\item{pcorTH}{Threshold for calling significant partial correlations. Default value is 0.1.
Usually, single-cell experiments profile large number of cells, so 0.1 is a 
sensible threshold.}
}
\value{
potencyInfer.l
A list incorporates the input list and a new list \code{potencyInfer.l}
contains sixteen objects:

potS
Inferred discrete potency states for each single cell. It is indexed so 
that the index increases as the signaling entropy of the state decreases

distPSPH
Table giving the distribution of single-cells across potency states and 
phenotypes

prob
Table giving the probabilities of each potency state per phenotype value

hetPS
The normalised Shannon Index of potency per phenotype value

cl
The co-expression clustering index for each single cell

pscl
The potency coexpression clustering label for each single cell

distPSCL
The distribution of single cell numbers per potency state and coexpression 
cluster

medLM
A matrix of medoids of gene expression for the selected landmarks

srPSCL
The average signaling entropy of single cells in each potency coexpression 
cluster

srLM
The average signaling entropy of single cells in each landmark

distPHLM
Table giving the distribution of single cell numbers per phenotype and 
landmark

cellLM
Nearest landmark for each single cell

cellLM2
A vector specifying the nearest and next-nearest landmark for each single 
cell

adj
Weighted adjacency matrix between landmarks with entries giving the number 
of single 
cells mapping closest to the two landmarks

pcorLM
Partial correlation matrix of landmarks as estimated from the expression 
medoids

netLM
Adjacency matrix of landmarks specifying which partial correlations are 
significant
}
\description{
This function infers the discrete potency states of single cells, 
its distribution across the single cell population, identifies 
potency-coexpression clusters of single cells, called landmarks, 
and finally infers the dependencies of these landmarks which can 
aid in recontructing lineage trajectories in time course scRNA-Seq 
experiments.
}
\details{
This function infers the discrete potency states of single cells, 
its distribution across the single cell population, identifies 
potency-coexpression clusters of single cells, called landmarks, 
and finally infers the dependencies of these landmarks which can 
aid in recontructing lineage trajectories in time course scRNA-Seq 
experiments.
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff AE, Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
