% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potS.v}
\alias{potS.v}
\title{Potency states for example plot scRNA-seq data}
\format{A numeric vector includes 3473 elements}
\usage{
data(potS.v)
}
\description{
Potency states infered from \code{PotencyInfer} function with 
example scRNA-seq data.
}
\details{
\itemize{
  \item Potency States : numbers indicates potency levels
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
\keyword{datasets}
