% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawExample.m}
\alias{rawExample.m}
\title{Raw example scRNA-seq data matrix}
\format{
A matrix with 19097 rows(genes) and 100 columns(cells)
}
\usage{
data(rawExample.m)
}
\description{
Raw example scRNA-seq data matrix of 50 human 
embryonic stem cells(hESC) and 50 mesoderm progenitor cells(EC)
from Chu et al. 2016.
}
\details{
\itemize{
  \item hESC : human embryonic stem cells(n=50)
  \item EC : mesoderm progenitor cells(n=50)
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Chu Li-Fang, Ning Leng, Jue Zhang, Zhonggang Hou, Daniel Mamott, 
David T. Vereide, Jeea Choi, Christina Kendziorski, Ron Stewart and James A. Thomson
\emph{Single-cell RNA-seq reveals novel regulators of human embryonic 
stem cell differentiation to definitive endoderm.}
Genome biology 17.1 (2016): 173.
doi:\href{https://doi.org/10.1186/s13059-016-1033-x}{
10.1186/s13059-016-1033-x}.
}
\keyword{datasets}
