% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SR.v}
\alias{SR.v}
\title{SR values for example scRNA-seq data}
\format{
A vector with 96 elements
}
\usage{
data(SR.v)
}
\description{
SR values calculated from \code{CompSRana} function with the input 
of example scRNA-seq data and network information.
}
\details{
\itemize{
  \item SR value : Normalized Signaling Entropy Rate
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
\keyword{datasets}
