% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsne.o}
\alias{tsne.o}
\title{Reduced dimension matrix for example plot scRNA-seq data}
\format{A numeric matrix with two colums}
\usage{
data(tsne.o)
}
\description{
Dimension reduced via \code{tsne}.
}
\details{
\itemize{
  \item data projection values in two-dimension space
}
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
\keyword{datasets}
