% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoDiffusionMap.R
\name{DoDiffusionMap}
\alias{DoDiffusionMap}
\title{Construct diffusion map and select root cell}
\usage{
DoDiffusionMap(Integration.l, mean_gap = 1, sd_gap = 1,
  root = c("cell", "state"), num_comp = 3, k = 30)
}
\arguments{
\item{Integration.l}{Typically, it is the output from \code{InferLandmark} function}

\item{mean_gap}{The threshold for HVG gene selection based on the mean expression
level. Default value is 1}

\item{sd_gap}{The threshold for HVG gene selection based on the expression
level's standard deviation. Default value is 1}

\item{root}{Define the root of the trajectory. \code{cell} indicates the
root is the cell with the highest SR value among all cells;
\code{state} indicates the root is the cell with the highest 
SR value within the highest inferred potency state}

\item{num_comp}{A numeric. The dimension of diffusion map, which is related 
to the following plot generation. Default is 3}

\item{k}{Number of nearest neighbors to consider. Default is 30}
}
\value{
An updated list objects, add three elements:

DM
A diffusion map object

DMEigen
The diffusion components of the diffusion map

root
The index of root cell for generating diffusion map plots
}
\description{
This function constructs diffusion map with highly variable
genes and visualize the result in 2D or 3D plots
}
\examples{
\dontrun{
data(Example.m)
data(net13Jun12.m)
Integration.l <- DoIntegPPI(exp.m = Example.m[, c(1:58,61:84,86:98,100)], ppiA.m = net13Jun12.m)
data(SR.v)
Integration.l$SR <- SR.v
InferPotency.o <- InferPotency(Integration.l)
DoDM.o <- DoDiffusionMap(InferPotency.o)}

}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.
}
